python-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
tiktoken==0.5.2
pypdf==3.17.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.26.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
